requirements.txt
Flask[async]==2.3.3
requests
pytube
PyMuPDF
pyngrok
elevenlabs
openai
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
langgraph==0.0.15
asgiref==3.7.2